requirements.txt
Pillow
lxml
rich
tqdm